+++ src/main/java/com/tienda/facil/core/controller/PedidoController.java
@@ -1,7 +1,7 @@
 package com.tienda.facil.core.controller;
 
 import com.tienda.facil.core.dto.request.PedidoDto;
-import com.tienda.facil.core.dto.ResponseDTO;
+import com.tienda.facil.core.dto.response.ResponseDTO;
 import com.tienda.facil.core.service.PedidoService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
